feat: vibrancy and root providers
