+++ src/renderer/src/App.tsx
@@ -1,7 +1,4 @@
-import { Provider } from "jotai"
 import { Outlet } from "react-router-dom"
-
-import { jotaiStore } from "./lib/jotai"
 
 function App() {
   return (
@@ -10,9 +7,8 @@ function App() {
         className="drag-region absolute inset-x-0 top-0 h-10 shrink-0"
         aria-hidden
       />
-      <Provider store={jotaiStore}>
-        <Outlet />
-      </Provider>
+
+      <Outlet />
     </>
   )
 }
