+++ src/renderer/src/pages/(main)/(context)/layout.tsx
@@ -4,11 +4,12 @@ import { Outlet } from "react-router-dom"
 export function Component() {
   return (
     <div className="flex h-full">
-      <div className="w-64 shrink-0 border-r bg-native pt-10">
+      <div className="w-64 shrink-0 border-r bg-native/80 pt-10 backdrop-blur">
         <FeedColumn />
       </div>
-
-      <Outlet />
+      <div className="flex flex-1 bg-background">
+        <Outlet />
+      </div>
     </div>
   )
 }
